mp4 hardcore

+++ api/views/query.py
@@ -454,8 +454,11 @@ class QueryUploadFilesView(APIView):
         thumb = resize_img(file, 256)
         if thumb:
             name_file_thumb, extension_thumb = os.path.splitext(thumb.name)
-            url_thumb = s3_upload_file(thumb, filename + '-thumb' + extension_thumb)
-            remove_file(thumb)
+            if extension_thumb != 'mp4':
+                url_thumb = s3_upload_file(thumb, filename + '-thumb' + extension_thumb)
+                remove_file(thumb)
+            else:
+                url_thumb = 'https://s3.amazonaws.com/linkup-photos/api/thumb-video-copy-thumb.jpg'
         else:
             url_thumb = ""
